hamieh-tunnel/pyproject.toml
[project]
name = "hamieh-tunnel"
version = "1.0.0"
description = "Production-grade SOCKS5 + TLS tunneling + VPN system with mobile API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Async HTTP + WebSocket
    "aiohttp>=3.9",
    "websockets>=12.0",
    # HTTP client (for verification + mobile API)
    "httpx[socks]>=0.27",
    # Config
    "pyyaml>=6.0",
    # CLI
    "click>=8.1",
    "rich>=13.0",
    # Security / crypto
    "cryptography>=42.0",
    "PyJWT>=2.8",
    # Rate limiting + caching
    "cachetools>=5.3",
    # Metrics
    "prometheus-client>=0.20",
    # Process management
    "psutil>=5.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "black>=24.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
hamieh = "cli.main:cli"
hamieh-server = "server.relay:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "tunnel", "client", "server", "mobile_api", "cli"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = false
